+++ crowdin/messages.json
@@ -571,6 +571,7 @@
   "790534726": "Implement controls to prevent unlawful activity",
   "793647899": "Your Binary.com affiliate account will be valid for Deriv.com as well. You will be able to earn commissions from your clients when they trade on either Binary.com or Deriv.com.",
   "795493132": "Poland",
+  "796127286": "10:20 - 11:20",
   "799457743": "Simple.",
   "799923297": "<0>Turnover</0><br/>Earn up to 1.5% commission for each contract bought by your clients. The lower the probability of winning, the higher your commission payout.",
   "800183478": "Security is important to us. We continuously improve our products and services by collaborating with independent security researchers worldwide.",
@@ -753,6 +754,7 @@
   "1023044581": "Deriv X dashboard with tools to track your online trading",
   "1023795011": "Even/Odd",
   "1023972728": "<0>Swap charge = volume × contract size × point value × swap rate</0>",
+  "1024070344": "Day3",
   "1024366221": "The page you are looking for does not exist.",
   "1024953822": "Trade forex, synthetics, and cryptocurrencies on the go",
   "1025032938": "Do I have to pay any fees to become a payment agent for Deriv?",
@@ -916,6 +918,7 @@
   "1238256430": "Parameters",
   "1240797431": "Start trading the world's most popular cryptocurrencies on Deriv, and take advantage of a highly volatile market with round-the-clock trading.",
   "1241426772": "<0>Margin = (volume × contract size × asset price) ÷ leverage</0>",
+  "1242374750": "Day2",
   "1244168183": "IBM",
   "1246145751": "Help centre | Frequently asked questions | Security | Deriv",
   "1248844726": "Deriv MT5 (DMT5)",
@@ -924,6 +927,7 @@
   "1254866713": "DMT5 trading dashboard",
   "1256295276": "Trading online can be exciting, but it’s important to be reminded that there are risks involved. We encourage all our users to secure their accounts and trade responsibly to experience the best in online trading.",
   "1256530668": "Go to your <0>Deriv X dashboard</0>.",
+  "1256749588": "Registrations 9:30 - 10:00",
   "1257029444": "Please check your email and click on the link provided to reset your password.",
   "1258668610": "Build your bot using drag and drop",
   "1259598687": "GBP/JPY",
@@ -999,6 +1003,7 @@
   "1362578283": "High",
   "1363124819": "You can cancel your withdrawal on the <0>Cashier page</0> by following the steps mentioned in the next paragraph. Please note that you cannot cancel withdrawals if:",
   "1363502727": "PnL for multipliers",
+  "1363510371": "Introduction - Overview of Brazil in Deriv",
   "1363645836": "Derived FX",
   "1364296672": "We will price and settle all contracts accurately and fairly, based on reliable data feeds and without undue delay, so that customers can trust our prices throughout the website.",
   "1364947573": "Dtrader artboard",
@@ -1030,6 +1035,7 @@
   "1400962248": "High-Close",
   "1403586449": "Competitively tighter spreads",
   "1403848854": "The Deriv MT5 Standard account offers new and experienced traders high leverage and variable spreads for maximum flexibility.",
+  "1404770708": "Brazil in numbers",
   "1405237333": "Head of People Management",
   "1406209174": "Trade with a regulated industry pioneer trusted by traders for more than 20 years.",
   "1406360927": "Minor pairs",
@@ -1239,6 +1245,7 @@
   "1637050346": "Why did my credit card deposit get declined?",
   "1637552314": "Once you are satisfied with the parameters that you have set, select either ‘Up’ or ‘Down’ to purchase your contract. Otherwise, continue to customise the parameters and place your order when you are satisfied with the conditions.",
   "1641578390": "1 }",
+  "1642229892": "Alcino",
   "1643887982": "Amazon.com Inc",
   "1644652204": "Fund your account with your preferred payment method.",
   "1644876351": "Go to portal",
@@ -1340,6 +1347,7 @@
   "1795749796": "Help centre | Frequently asked questions | DBot | Deriv",
   "1798765888": "Copy the link in the <0>Landing page URL</0> field. This is the referral link that you can share with your clients.",
   "1799377913": "Option trades available on {{market_name}}",
+  "1799844411": " Sebastian",
   "1802409084": "Pip value",
   "1804642811": "Create demo account",
   "1805914152": "Fiat onramp - Buy crypto on popular exchanges.",
@@ -1804,6 +1812,10 @@
   "-1961258037": "Didn't receive an email from us? Here's what could've happened.",
   "-276906320": "If you don't see an email from us within a few minutes, a few things could have happened:",
   "-1878043830": "Re-enter your email and try again",
+  "-754682908": "Day1",
+  "-1929602503": "November",
+  "-1824897170": "10:00 - 10:20",
+  "-1965789818": "Break 11:00 - 11:15",
   "-1555139667": "Schedule",
   "-289882353": "Day 2",
   "-1715499111": "Day 3",
