+++ src/pages/conference/timeline/_data-schedule.tsx
@@ -1,0 +1,356 @@
+import React from 'react'
+import { Localize } from 'components/localization'
+
+const events_data = [
+    {
+        id: 'day_01',
+        title: <Localize translate_text="Day1" />,
+        date: <Localize translate_text="November" />,
+        period: <Localize translate_text="Registrations 9:30 - 10:00" />,
+        data: [
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+        ],
+    },
+
+    {
+        id: 'day_02',
+        title: <Localize translate_text="Day2" />,
+        date: <Localize translate_text="November" />,
+        period: <Localize translate_text="Registrations 9:30 - 10:00" />,
+        data: [
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+        ],
+    },
+    {
+        id: 'day_03',
+        title: <Localize translate_text="Day3" />,
+        date: <Localize translate_text="November" />,
+        period: <Localize translate_text="Registrations 9:30 - 10:00" />,
+        data: [
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+            {
+                timings: [
+                    {
+                        time: <Localize translate_text="10:00 - 10:20" />,
+                        id: 'time_01',
+                    },
+                    {
+                        time: <Localize translate_text="10:20 - 11:20" />,
+                        id: 'time_02',
+                    },
+                ],
+                speakers: [
+                    {
+                        id: 'name_01',
+                        name: <Localize translate_text="Alcino" />,
+                        topic: (
+                            <Localize translate_text="Introduction - Overview of Brazil in Deriv" />
+                        ),
+                    },
+                    {
+                        id: 'name_02',
+                        name: <Localize translate_text=" Sebastian" />,
+                        topic: <Localize translate_text="Brazil in numbers" />,
+                    },
+                ],
+                free_time: <Localize translate_text="Break 11:00 - 11:15" />,
+            },
+        ],
+    },
+]
+
+export default events_data
